+++ tests/lib.rs
@@ -1,5 +1,5 @@
-use std::fs::File;
 use std::path::Path;
+use std::{fs::File, path::PathBuf};
 use tiled::{
     error::TiledError, layers::LayerData, map::Map, properties::PropertyValue, tileset::Tileset,
 };
@@ -11,11 +11,11 @@ fn parse_map_without_source(p: impl AsRef<Path>) -> Result<Map, TiledError> {
 
 #[test]
 fn test_gzip_and_zlib_encoded_and_raw_are_the_same() {
-    let z = parse_map_without_source("assets/tiled_base64_zlib.tmx").unwrap();
-    let g = parse_map_without_source("assets/tiled_base64_gzip.tmx").unwrap();
-    let r = parse_map_without_source("assets/tiled_base64.tmx").unwrap();
-    let zstd = parse_map_without_source("assets/tiled_base64_zstandard.tmx").unwrap();
-    let c = parse_map_without_source("assets/tiled_csv.tmx").unwrap();
+    let z = Map::parse_file("assets/tiled_base64_zlib.tmx").unwrap();
+    let g = Map::parse_file("assets/tiled_base64_gzip.tmx").unwrap();
+    let r = Map::parse_file("assets/tiled_base64.tmx").unwrap();
+    let zstd = Map::parse_file("assets/tiled_base64_zstandard.tmx").unwrap();
+    let c = Map::parse_file("assets/tiled_csv.tmx").unwrap();
     assert_eq!(z, g);
     assert_eq!(z, r);
     assert_eq!(z, c);
@@ -37,17 +37,30 @@ fn test_gzip_and_zlib_encoded_and_raw_are_the_same() {
 
 #[test]
 fn test_external_tileset() {
-    let r = parse_map_without_source("assets/tiled_base64.tmx").unwrap();
+    let r = Map::parse_file("assets/tiled_base64.tmx").unwrap();
     let mut e = Map::parse_file("assets/tiled_base64_external.tmx").unwrap();
-    e.source = None;
-    // Compare everything BUT source
+    e.tilesets[0].source = None;
     assert_eq!(r, e);
 }
 
 #[test]
+fn test_sources() {
+    let e = Map::parse_file("assets/tiled_base64_external.tmx").unwrap();
+    assert_eq!(
+        e.tilesets[0].source,
+        Some(PathBuf::from("assets/tilesheet.tsx"))
+    );
+    assert_eq!(
+        e.tilesets[0].image.as_ref().unwrap().source,
+        PathBuf::from("assets/tilesheet.png")
+    );
+}
+
+#[test]
 fn test_just_tileset() {
-    let r = parse_map_without_source("assets/tiled_base64.tmx").unwrap();
-    let t = Tileset::parse(File::open(Path::new("assets/tilesheet.tsx")).unwrap(), 1).unwrap();
+    let r = Map::parse_file("assets/tiled_base64_external.tmx").unwrap();
+    let path = "assets/tilesheet.tsx";
+    let t = Tileset::parse_with_path(File::open(path).unwrap(), 1, path).unwrap();
     assert_eq!(r.tilesets[0], t);
 }
 
@@ -70,7 +83,7 @@ fn test_infinite_tileset() {
 
 #[test]
 fn test_image_layers() {
-    let r = parse_map_without_source("assets/tiled_image_layers.tmx").unwrap();
+    let r = Map::parse_file("assets/tiled_image_layers.tmx").unwrap();
     assert_eq!(r.image_layers.len(), 2);
     {
         let first = &r.image_layers[0];
@@ -88,7 +101,7 @@ fn test_image_layers() {
             .image
             .as_ref()
             .expect(&format!("{}'s image shouldn't be None", second.name));
-        assert_eq!(image.source, "tilesheet.png");
+        assert_eq!(image.source, PathBuf::from("assets/tilesheet.png"));
         assert_eq!(image.width, 448);
         assert_eq!(image.height, 192);
     }
@@ -96,7 +109,7 @@ fn test_image_layers() {
 
 #[test]
 fn test_tile_property() {
-    let r = parse_map_without_source("assets/tiled_base64.tmx").unwrap();
+    let r = Map::parse_file("assets/tiled_base64.tmx").unwrap();
     let prop_value: String = if let Some(&PropertyValue::StringValue(ref v)) =
         r.tilesets[0].tiles[0].properties.get("a tile property")
     {
@@ -109,7 +122,7 @@ fn test_tile_property() {
 
 #[test]
 fn test_layer_property() {
-    let r = parse_map_without_source(&Path::new("assets/tiled_base64.tmx")).unwrap();
+    let r = Map::parse_file(&Path::new("assets/tiled_base64.tmx")).unwrap();
     let prop_value: String =
         if let Some(&PropertyValue::StringValue(ref v)) = r.layers[0].properties.get("prop3") {
             v.clone()
@@ -121,7 +134,7 @@ fn test_layer_property() {
 
 #[test]
 fn test_object_group_property() {
-    let r = parse_map_without_source("assets/tiled_object_groups.tmx").unwrap();
+    let r = Map::parse_file("assets/tiled_object_groups.tmx").unwrap();
     let prop_value: bool = if let Some(&PropertyValue::BoolValue(ref v)) = r.object_groups[0]
         .properties
         .get("an object group property")
@@ -134,7 +147,7 @@ fn test_object_group_property() {
 }
 #[test]
 fn test_tileset_property() {
-    let r = parse_map_without_source("assets/tiled_base64.tmx").unwrap();
+    let r = Map::parse_file("assets/tiled_base64.tmx").unwrap();
     let prop_value: String = if let Some(&PropertyValue::StringValue(ref v)) =
         r.tilesets[0].properties.get("tileset property")
     {